WIP: Build conda packages for k2.
