+++ k2/csrc/host/CMakeLists.txt
@@ -25,7 +25,11 @@ set_target_properties(fsa PROPERTIES OUTPUT_NAME "k2fsa")
 target_link_libraries(fsa PUBLIC k2_log)
 target_link_libraries(fsa PUBLIC k2_nvtx)
 target_include_directories(fsa PUBLIC ${CUDA_TOOLKIT_INCLUDE})
-target_link_libraries(fsa PUBLIC -L${CUDA_TOOLKIT_ROOT_DIR}/targets/x86_64-linux/lib nvToolsExt)
+target_link_libraries(fsa
+  PUBLIC
+    -L${CUDA_TOOLKIT_ROOT_DIR}/targets/x86_64-linux/lib  # for /usr/local/cuda
+    -L${CUDA_TOOLKIT_ROOT_DIR}/lib  # for conda
+  nvToolsExt)
 
 #---------------------------- Test K2 host sources ----------------------------
